requirements.txt
python-dotenv
langchain
langchain_openai
langchain_core